requirements.txt
torch>=2.0.0
transformers>=4.30.0
tensorrt-llm>=0.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.12.0
tqdm>=4.65.0
pynvml>=11.5.0
psutil>=5.9.0
pyyaml>=6.0.0
pytest>=7.0.0
jupyter>=1.0.0
ipykernel>=6.0.0 